refactor; setup pull app
