+++ cognite_toolkit/_cdf_tk/apps/__init__.py
@@ -2,7 +2,8 @@ from ._auth_app import AuthApp
 from ._core_app import CoreApp
 from ._landing_app import LandingApp
 from ._modules_app import ModulesApp
+from ._pull_app import PullApp
 from ._repo_app import RepoApp
 from ._run import RunApp
 
-__all__ = ["CoreApp", "AuthApp", "LandingApp", "ModulesApp", "RunApp", "RepoApp"]
+__all__ = ["CoreApp", "AuthApp", "LandingApp", "ModulesApp", "RunApp", "RepoApp", "PullApp"]
